requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pystray>=0.19.0
Pillow>=9.0.0 